22/02/2023 - Update 002
added container functionality (currently just debug logs, no actual looting is in place). I also added two placeholder models for the ammo and consumable containers, and got the loot generator in place! lots of work done today; well done me!

+++ Assets/Scripts/Entities/TestDummy.cs
@@ -8,7 +8,7 @@ namespace com.limphus.retro_survival_shooter
     {
         [SerializeField] private double health;
 
-        public void Damage(double amount)
+        public void Damage(float amount)
         {
             health -= amount;
